requirements.txt
cyrtranslit>=0.4
discord.py>=1.5.1
numpy>=1.18.5
scipy>=1.5.4
tensorflow>=2.3.1
tqdm>=4.51.0
requests>=2.24.0
